pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel>=0.42.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langquality"
version = "1.0.0"
description = "Language Quality Toolkit for Low-Resource Languages"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "LangQuality Community", email = "frejus.laleye@gmail.com"}
]
maintainers = [
    {name = "LangQuality Community", email = "frejus.laleye@gmail.com"}
]
keywords = [
    "nlp",
    "low-resource-languages",
    "data-quality",
    "linguistics",
    "language-packs",
    "text-analysis",
    "dataset-quality",
    "multilingual",
    "language-toolkit",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Natural Language :: English",
    "Natural Language :: French",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "spacy>=3.5.0",
    "textstat>=0.7.3",
    "nltk>=3.8",
    "chardet>=5.2.0",
    "plotly>=5.14.0",
    "jinja2>=3.1.2",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "reportlab>=4.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-benchmark>=4.0.0",
    "pytest-mock>=3.11.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=23.12.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "bandit>=1.7.6",
    "pre-commit>=3.6.0",
    "ipython>=8.12.0",
    "ipdb>=0.13.13",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]
build = [
    "build>=1.0.0",
    "twine>=4.0.0",
    "setuptools>=68.0.0",
    "wheel>=0.42.0",
]
spacy = [
    "spacy>=3.5.0",
]
nltk = [
    "nltk>=3.8",
]

[project.urls]
Homepage = "https://github.com/laleye/langquality"
Documentation = "https://langquality.readthedocs.io"
Repository = "https://github.com/laleye/langquality"
"Bug Tracker" = "https://github.com/laleye/langquality/issues"
Changelog = "https://github.com/laleye/langquality/blob/main/CHANGELOG.md"
Discussions = "https://github.com/laleye/langquality/discussions"

[project.scripts]
langquality = "langquality.__main__:main"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["langquality*"]
namespaces = false

[tool.setuptools.package-data]
langquality = [
    "resources/*",
    "outputs/templates/*",
    "language_packs/packs/*/config.yaml",
    "language_packs/packs/*/metadata.json",
    "language_packs/packs/*/resources/*",
    "language_packs/packs/*/README.md",
    "language_packs/packs/_template/**/*",
]
fongbe_quality = [
    "resources/*",
    "outputs/templates/*",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/venv/*", "*/.venv/*"]

[tool.coverage.run]
source = ["src/langquality", "src/fongbe_quality"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/migrations/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests", "docs", ".venv", "venv", "build", "dist"]
skips = ["B101", "B601"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "numpy.*",
    "spacy.*",
    "nltk.*",
    "plotly.*",
    "reportlab.*",
]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    "venv",
    ".tox",
    ".mypy_cache",
    ".pytest_cache",
]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:D100,D101,D102,D103,D104",
]
docstring-convention = "google"
